Update regarding feedbacks
Hugo, I updated the readme to replace Mango Pay by MangoPay :) 

Thanks for the feedbacks on good practices with Composer. I'm a newbie with all that kind of stuff so I was expecting that someone more experienced, do this job but ... ;) 

Regarding samples,I agree but there was already some in the readme.